Warning tidy-up for FreeBSD.
- if dlfunc is available use that. Allows -Werror to be re-enabled.
- use modern struct member initializers
- cope with FD_SETSIZE being explicitly unsigned
- remove stray brackets
- remove unused objects
- use %ju to print sig_atomic_t, it might be bigger than int

+++ lib/event.c
@@ -311,7 +311,8 @@ int ev_fd(ev_source *ev,
 
   D(("registering %s fd %d callback %p %p", modenames[mode], fd,
      (void *)callback, u));
-  if(fd >= FD_SETSIZE)
+  /* FreeBSD defines FD_SETSIZE as 1024u for some reason */
+  if((unsigned)fd >= FD_SETSIZE)
     return -1;
   assert(mode < ev_nmodes);
   if(ev->mode[mode].nfds >= ev->mode[mode].fdslots) {
